Move windows threading impl into the header-file
