+++ include/core/platform/lf_windows_support.h
@@ -27,7 +27,13 @@ THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 /** Windows API support for the C target of Lingua Franca.
  *
  *  @author{Soroush Bateni <soroush@utdallas.edu>}
- *
+ *  @author{Erling Jellum <erling.r.jellum@ntnu.no>}
+ *  
+ * The API is implemented in the header files. This is also the case for Linux
+ * and macos. This is to enable having both the unthreaded and the threaded API
+ * available in the same program. This is needed for unthreaded programs with 
+ * tracing. trace.c can then do #define LF_TRACING and then include this file
+ *  
  * All functions return 0 on success.
  *
  * @see https://gist.github.com/Soroosh129/127d1893fa4c1da6d3e1db33381bb273
@@ -43,8 +49,6 @@ THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #if defined LF_THREADED || defined LF_TRACING
 #if __STDC_VERSION__ < 201112L || defined (__STDC_NO_THREADS__) // (Not C++11 or later) or no threads support
-
-
 /**
  * On Windows, one could use both a mutex or
  * a critical section for the same purpose. However,
@@ -61,6 +65,179 @@ typedef _lf_mutex_t _lf_critical_section_t;
 typedef CONDITION_VARIABLE _lf_cond_t;
 typedef HANDLE _lf_thread_t;
 
+/**
+ * Create a new thread, starting with execution of lf_thread
+ * getting passed arguments. The new handle is stored in thread.
+ *
+ * @return 0 on success, errno otherwise.
+ */
+int lf_thread_create(_lf_thread_t* thread, void *(*lf_thread) (void *), void* arguments) {
+    uintptr_t handle = _beginthreadex(NULL, 0, lf_thread, arguments, 0, NULL);
+    *thread = (HANDLE)handle;
+    if(handle == 0){
+        return errno;
+    }else{
+        return 0;
+    }
+}
+
+/**
+ * Make calling thread wait for termination of the thread.  The
+ * exit status of the thread is stored in thread_return, if thread_return
+ * is not NULL.
+ *
+ * @return 0 on success, EINVAL otherwise.
+ */
+int lf_thread_join(_lf_thread_t thread, void** thread_return) {
+    DWORD retvalue = WaitForSingleObject(thread, INFINITE);
+    if(retvalue == WAIT_FAILED){
+        return EINVAL;
+    }
+    return 0;
+}
+
+/**
+ * Initialize a critical section.
+ *
+ * @return 0 on success, 1 otherwise.
+ */
+int lf_mutex_init(_lf_critical_section_t* critical_section) {
+    // Set up a recursive mutex
+    InitializeCriticalSection((PCRITICAL_SECTION)critical_section);
+    if(critical_section != NULL){
+        return 0;
+    }else{
+        return 1;
+    }
+}
+
+/**
+ * Lock a critical section.
+ *
+ * From https://docs.microsoft.com/en-us/windows/win32/api/synchapi/nf-synchapi-entercriticalsection:
+ *    "This function can raise EXCEPTION_POSSIBLE_DEADLOCK if a wait operation on the critical section times out.
+ *     The timeout interval is specified by the following registry value:
+ *     HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Control\Session Manager\CriticalSectionTimeout.
+ *     Do not handle a possible deadlock exception; instead, debug the application."
+ *
+ * @return 0
+ */
+int lf_mutex_lock(_lf_critical_section_t* critical_section) {
+    // The following Windows API does not return a value. It can
+    // raise a EXCEPTION_POSSIBLE_DEADLOCK. See synchapi.h.
+    EnterCriticalSection((PCRITICAL_SECTION)critical_section);
+    return 0;
+}
+
+/**
+ * Leave a critical_section.
+ *
+ * @return 0
+ */
+int lf_mutex_unlock(_lf_critical_section_t* critical_section) {
+    // The following Windows API does not return a value.
+    LeaveCriticalSection((PCRITICAL_SECTION)critical_section);
+    return 0;
+}
+
+/**
+ * Initialize a conditional variable.
+ *
+ * @return 0
+ */
+int lf_cond_init(_lf_cond_t* cond) {
+    // The following Windows API does not return a value.
+    InitializeConditionVariable((PCONDITION_VARIABLE)cond);
+    return 0;
+}
+
+/**
+ * Wake up all threads waiting for condition variable cond.
+ *
+ * @return 0
+ */
+int lf_cond_broadcast(_lf_cond_t* cond) {
+    // The following Windows API does not return a value.
+    WakeAllConditionVariable((PCONDITION_VARIABLE)cond);
+    return 0;
+}
+
+/**
+ * Wake up one thread waiting for condition variable cond.
+ *
+ * @return 0
+ */
+int lf_cond_signal(_lf_cond_t* cond) {
+    // The following Windows API does not return a value.
+    WakeConditionVariable((PCONDITION_VARIABLE)cond);
+    return 0;
+}
+
+/**
+ * Wait for condition variable "cond" to be signaled or broadcast.
+ * "mutex" is assumed to be locked before.
+ *
+ * @return 0 on success, 1 otherwise.
+ */
+int lf_cond_wait(_lf_cond_t* cond, _lf_critical_section_t* critical_section) {
+    // According to synchapi.h, the following Windows API returns 0 on failure,
+    // and non-zero on success.
+    int return_value =
+     (int)SleepConditionVariableCS(
+         (PCONDITION_VARIABLE)cond,
+         (PCRITICAL_SECTION)critical_section,
+         INFINITE
+     );
+     switch (return_value) {
+        case 0:
+            // Error
+            return 1;
+            break;
+
+        default:
+            // Success
+            return 0;
+            break;
+     }
+}
+
+/**
+ * Block current thread on the condition variable until condition variable
+ * pointed by "cond" is signaled or time pointed by "absolute_time_ns" in
+ * nanoseconds is reached.
+ *
+ * @return 0 on success and LF_TIMEOUT on timeout, 1 otherwise.
+ */
+int lf_cond_timedwait(_lf_cond_t* cond, _lf_critical_section_t* critical_section, instant_t absolute_time_ns) {
+    // Convert the absolute time to a relative time
+    instant_t current_time_ns;
+    lf_clock_gettime(&current_time_ns);
+    interval_t relative_time_ns = (absolute_time_ns - current_time_ns);
+    if (relative_time_ns <= 0) {
+      // physical time has already caught up sufficiently and we do not need to wait anymore
+      return 0;
+    }
+
+    // convert ns to ms and round up to closest full integer
+    DWORD relative_time_ms = (relative_time_ns + 999999LL) / 1000000LL;
+
+    int return_value =
+     (int)SleepConditionVariableCS(
+         (PCONDITION_VARIABLE)cond,
+         (PCRITICAL_SECTION)critical_section,
+         relative_time_ms
+     );
+    if (return_value == 0) {
+      // Error
+      if (GetLastError() == ERROR_TIMEOUT) {
+        return _LF_TIMEOUT;
+      }
+      return 1;
+    }
+
+    // Success
+    return 0;
+}
 #else
 #include "lf_C11_threads_support.h"
 #endif
